Updating tasks to actually work well ;)
git-svn-id: http://svn.igniterealtime.org/svn/repos/spark/trunk@8899 b35dd754-fafc-0310-a699-88a17e54d16e

+++ src/java/org/jivesoftware/sparkimpl/plugin/scratchpad/Tasks.java
@@ -144,12 +144,12 @@ public class Tasks implements PrivateData {
 
             if (eventType == XmlPullParser.START_TAG && "dueDate".equals(parser.getName())) {
                 String dueDate = parser.nextText();
-                task.setDueDate(Integer.parseInt(dueDate));
+                task.setDueDate(Long.parseLong(dueDate));
             }
 
             if (eventType == XmlPullParser.START_TAG && "creationDate".equals(parser.getName())) {
                 String creationDate = parser.nextText();
-                task.setCreatedDate(Integer.parseInt(creationDate));
+                task.setCreatedDate(Long.parseLong(creationDate));
             }
 
             if (eventType == XmlPullParser.START_TAG && "completed".equals(parser.getName())) {
